Remove the dead video call, thread request, and upstream channel read DTOs
